EDGE/requirements copy.txt
paho-mqtt>=1.6.1
numpy>=1.21.0
amqtt>=0.11.0
netifaces>=0.11.0

# GUI (optional - for visualization)
PyQt6>=6.7.0
pyqtgraph>=0.13.3

